feat(comment): add model for comment learning material and assignment
