+++ src/models/class-based-learning/assignment/assignmentComment.model.ts
@@ -1,19 +1,21 @@
-import { usersTable } from '@/models';
-import { integer, pgTable, serial, text, timestamp } from 'drizzle-orm/pg-core';
-import { assignmentSubmissionsTable } from './assignmentSubmission.model';
+import { integer, pgTable, serial } from 'drizzle-orm/pg-core';
+import { commentsTable } from '@/models/comment/comment.model';
 import { assignmentsTable } from './assignment.model';
 
+/**
+ * AssignmentComment table - for public comments on assignments
+ * This table links comments to assignments (visible to everyone in the class)
+ * Private comments on submissions are handled separately (not implemented yet)
+ */
 export const assignmentCommentsTable = pgTable('assignment_comments', {
-    commentId: serial('comment_id').primaryKey(),
-    // assignmentId serves public comment feature, but is not planned to be implemented at the present
-    assignmentId: integer('assignment_id').references(() => assignmentsTable.assignmentId, { onDelete: 'cascade' }),
-    submissionId: integer('submission_id')
+    assignmentCommentId: serial('assignment_comment_id').primaryKey(),
+    commentId: integer('comment_id')
         .notNull()
-        .references(() => assignmentSubmissionsTable.submissionId, { onDelete: 'cascade' }),
-    senderId: integer('sender_id')
+        .references(() => commentsTable.commentId, { onDelete: 'cascade' }),
+    assignmentId: integer('assignment_id')
         .notNull()
-        .references(() => usersTable.userId, { onDelete: 'cascade' }),
-    content: text('content').notNull().default(''),
-    createdAt: timestamp('created_at', { withTimezone: true }).notNull().defaultNow(),
-    updatedAt: timestamp('updated_at', { withTimezone: true }),
+        .references(() => assignmentsTable.assignmentId, { onDelete: 'cascade' }),
 });
+
+export type TypeSelectAssignmentComment = typeof assignmentCommentsTable.$inferSelect;
+export type TypeInsertAssignmentComment = typeof assignmentCommentsTable.$inferInsert;
