+++ src/models/class-based-learning/assignment/index.ts
@@ -5,3 +5,5 @@ export * from '../assignment/assignmentAttachments.model';
 export * from '../assignment/assignmentSubmission.model';
 
 export * from '../assignment/assignmentSubmissionAttachments.model';
+
+export * from '../assignment/assignmentComment.model';
